Separate derive helpers into a separate crate
