+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
   "migrator-tests",
   "proc-macros-helpers",
   "derive",
+  "derive_helpers",
   "derive_tests",
   "models",
   "orm-tests",
@@ -29,6 +30,7 @@ documentation = "https://codebreather.com/oyelowo"
 
 [workspace.dependencies]
 surreal-derive = { path = "derive" }
+surreal-derive-helpers = { path = "derive_helpers" }
 surreal-orm = { path = "orm" }
 surreal-query-builder = { path = "query-builder" }
 proc-macros-helpers = { path = "proc-macros-helpers" }
