+++ derive_helpers/src/lib.rs
@@ -1,0 +1,10 @@
+/*
+ * Author: Oyelowo Oyedayo
+ * Email: oyelowo.oss@gmail.com
+ * Copyright (c) 2023 Oyelowo Oyedayo
+ * Licensed under the MIT license
+ */
+
+pub mod migrations;
+pub mod models;
+
