+++ derive_helpers/src/models/field_meta/type_stripper.rs
