+++ derive_helpers/src/models/mod.rs
@@ -5,10 +5,29 @@
  * Licensed under the MIT license
  */
 
-pub mod edge;
-pub mod node;
-pub mod object;
+pub(crate) mod casing;
+pub(crate) mod edge;
+pub(crate) mod errors;
+pub(crate) mod field_meta;
+pub(crate) mod keywords;
+pub(crate) mod node;
+pub(crate) mod object;
+pub(crate) mod table_meta;
+pub(crate) mod token_codegen;
+pub(crate) mod utils;
+pub(crate) mod variables;
 
+pub use casing::*;
+pub use errors::*;
+pub use field_meta::*;
+pub use keywords::*;
+pub use table_meta::*;
+pub use token_codegen::*;
+pub use utils::*;
+pub use variables::*;
+pub use edge::*;
+pub use node::*;
+pub use object::*;
 
 #[derive(Debug, Copy, Clone)]
 pub enum DataType {
