+++ derive_helpers/src/models/token_codegen/link_methods.rs
