+++ derive_helpers/src/models/token_codegen/simple.rs
