+++ derive_helpers/src/models/token_codegen/struct_partial.rs
